requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
Django==5.1
django-appconf==1.0.6
django-cleanup==8.1.0
django-compressor==4.5.1
django-cors-headers==4.3.1
django-js-asset==2.2.0
django-taggit==5.0.1
django-tailwind==3.8.0
gunicorn==21.2.0
idna==3.8
iyzipay==1.0.42
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.0
pillow==10.2.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
rcssmin==1.1.2
requests==2.32.3
rich==13.8.0
rjsmin==1.2.2
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.9.0.20240821
tzdata==2024.1
urllib3==2.2.2
